refactor: Bookmarker 로직 수정
