+++ package.json
@@ -11,7 +11,7 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.10",
     "axios": "^1.3.3",
-    "klinecharts": "^9.1.0",
+    "klinecharts": "^9.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-query": "^3.39.3",
