+++ src/components/CoinDetail/DetailNameBox.tsx
@@ -1,7 +1,7 @@
 import BookMarker from "components/Bookmarker/BookMarker";
 import React from "react";
 import styled from "styled-components";
-import { market } from "Types/types";
+import { market } from "types/types";
 
 interface NameBoxProps {
   select: string;
@@ -16,7 +16,7 @@ const DetailNameBox = ({ select, focus }: NameBoxProps) => {
         <p>{focus?.korean_name}</p>
         <h4>{focus?.market}</h4>
       </div>
-      <BookMarker focus={focus} />
+      <BookMarker />
     </NameBox>
   );
 };
